+++ Papeleria/src/main/java/co/edu/unbosque/Papeleria/interfacesService/DetalleCompraCRUD.java
@@ -8,8 +8,8 @@ import co.edu.unbosque.Papeleria.modelo.DetalleVentas;
 
 public interface DetalleCompraCRUD {
 	public List<DetalleCompra> listBuyRep();
-	public Optional<DetalleCompra> searchBuyRep(Long id);
-	public String deleteBuyRep(Long id);
+	public Optional<DetalleCompra> searchBuyRep(Integer id);
+	public String deleteBuyRep(Integer id);
 	public DetalleCompra insertBuyRep(DetalleCompra BuyRep);
 	public DetalleCompra editBuyRep(DetalleCompra BuyRep);
